chore: improve the stability of unit test
